requirements.txt
openai
pandas
smtplib
python-dotenv